+++ Potestas/Potestas.sln
@@ -1,13 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.28803.452
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.168
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Potestas", "Potestas\Potestas.csproj", "{EEB46883-CBA0-4919-B984-A3254930E9CA}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Potestas.Tests", "Potestas.Tests\Potestas.Tests.csproj", "{478AA982-EC6E-43D8-A24A-F7BE3C4E0AF0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Potestas.Apps.Terminal", "Potestas.Apps.Terminal\Potestas.Apps.Terminal.csproj", "{3C3E81BE-F2C2-406D-B75D-599B6527F377}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Potestas.Apps.Terminal", "Potestas.Apps.Terminal\Potestas.Apps.Terminal.csproj", "{3C3E81BE-F2C2-406D-B75D-599B6527F377}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
