+++ Potestas/Potestas/ApplicationFrame/SourceRegistration/SourceStatus.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Potestas.ApplicationFrame.SourceRegistration
+{
+    public enum SourceStatus
+    {
+        Pending,
+        Running,
+        Completed,
+        Failed
+    }
+}
