+++ Potestas/Potestas/Factories/first attempt/AnalizerFactories/LINQAnalizerFactory.cs
@@ -1,0 +1,12 @@
+﻿//using Potestas.Analizers;
+
+//namespace Potestas.Factories.AnalizerFactories
+//{
+//    public class LINQAnalizerFactory : IAnalizerFactory<IEnergyObservation>
+//    {
+//        public IEnergyObservationAnalizer CreateAnalizer(IEnergyObservationStorage<IEnergyObservation> observationStorage)
+//        {
+//            return new LINQAnalizer(observationStorage);
+//        }
+//    }
+//}
