+++ Potestas/Potestas/Factories/first attempt/SerializerFactories/EnergyObservationJsonSerializerFactory.cs
@@ -1,0 +1,12 @@
+﻿//using Potestas.Serializers;
+
+//namespace Potestas.Factories.SerializerFactories
+//{
+//    public class EnergyObservationJsonSerializerFactory : ISerializerFactory<IEnergyObservation>
+//    {
+//        public ISerializer<IEnergyObservation> CreateSerializer()
+//        {
+//            return new JsonSerializer<IEnergyObservation>();
+//        }
+//    }
+//}
