+++ Potestas/Potestas/Factories/first attempt/StreamProcessingFactories/SerializeProcessorFactory.cs
@@ -1,0 +1,12 @@
+﻿//using Potestas.Processors;
+
+//namespace Potestas.Factories.StreamProcessingFactories
+//{
+//    public class SerializeProcessorFactory : IStreamProcessingFactory<IEnergyObservation>
+//    {
+//        public IStreamProcessor<IEnergyObservation> CreateStreamProcessor()
+//        {
+//            return new SerializeProcessor<IEnergyObservation>();
+//        }
+//    }
+//}
